Clean up home page
