+++ website/package.json
@@ -25,7 +25,7 @@
   },
   "dependencies": {
     "@docusaurus/core": "^3.0.0",
-    "@vis.gl/docusaurus-website": "1.0.0-alpha.19",
+    "@vis.gl/docusaurus-website": "1.0.0-alpha.21",
     "clsx": "^1.1.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
